refactor to withprogress and update tests
